% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_text.r
\name{create_cv_folds}
\alias{create_cv_folds}
\title{Create 10-fold cross-validation samples}
\usage{
create_cv_folds(train_x_class, train_y_class, category)
}
\arguments{
\item{train_x_class}{Training data for predictors}

\item{train_y_class}{Training data for outcomes}

\item{category}{The target binary category.}
}
\value{
10-fold cross-validation samples
}
\description{
Create 10-fold cross-validation samples
}
