% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_text.r
\name{create_search_spaces}
\alias{create_search_spaces}
\title{Create search spaces for the algorithms based on the hyperparameters}
\usage{
create_search_spaces(train_x_class, category, lasso_sepc, rand_spec, xg_spec)
}
\arguments{
\item{train_x_class}{Training data for predictors}

\item{category}{The target binary category}

\item{rand_spec}{The tuning parameters for random forest}

\item{xg_spec}{The tuning parameters for XGBoost}

\item{lasso_spec}{The tuning parameters for lasso}
}
\value{
A list of the search spaces for lasso, random forest, and XGBoost.
}
\description{
Create search spaces for the algorithms based on the hyperparameters
}
