% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_text.r
\name{create_tunes}
\alias{create_tunes}
\title{Create tuning parameters for algorithms (i.e., lasso, random forest, and XGBoost).}
\usage{
create_tunes(mode = "classification")
}
\arguments{
\item{mode}{The mode for the model specification.}
}
\value{
A list output that contains the tuning parameters for lasso, random forest, and XGBoost.
}
\description{
Create tuning parameters for algorithms (i.e., lasso, random forest, and XGBoost).
}
