% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_text.r
\name{create_workflows}
\alias{create_workflows}
\title{Create workflows for the algorithms based on the hyperparameters}
\usage{
create_workflows(lasso_spec, rand_spec, xg_spec, category)
}
\arguments{
\item{lasso_spec}{The model specification for lasso}

\item{rand_spec}{The model specification for random forest}

\item{xg_spec}{The model specification for XGBoost}

\item{category}{The target binary category}
}
\value{
A list of the workflows for lasso, random forest, and XGBoost.
}
\description{
Create workflows for the algorithms based on the hyperparameters
}
