% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{evaluate_class_fit}
\alias{evaluate_class_fit}
\title{Evaluate classification model outputs}
\usage{
evaluate_class_fit(model)
}
\arguments{
\item{model}{classification model outputs}

\item{test_y_class}{Predictors of the test dataset}

\item{test_x_class}{Outcomes of the test dataset}
}
\value{
A dataframe of two columns (truth, estimate)
}
\description{
Evaluate classification model outputs
}
