% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{evaluate_reg_fit}
\alias{evaluate_reg_fit}
\title{Evaluate regression model outputs}
\usage{
evaluate_reg_fit(model)
}
\arguments{
\item{model}{regression model outputs}
}
\value{
A dataframe of two columns (truth, estimate)
}
\description{
Evaluate regression model outputs
}
