% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_text.r
\name{select_best}
\alias{select_best}
\title{Select the best output for each algorithm based on the hyperparameters and 10-fold cross-validation.}
\usage{
select_best(lasso_wf, rand_wf, xg_wf)
}
\arguments{
\item{lasso_wf}{A lasso workflow (including the search space for the model)}

\item{rand_wf}{A random forest workflow (including the search space for the model)}

\item{xg_wf}{An XGBoost workflow (including the search space for the model)}
}
\value{
A list output that contains the best model output for lasso, random forest, and XGBoost.
}
\description{
Select the best output for each algorithm based on the hyperparameters and 10-fold cross-validation.
}
