% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{topn_vip}
\alias{topn_vip}
\title{Visualize the importance of top 20 features}
\usage{
topn_vip(model)
}
\arguments{
\item{model_outputs}{model outputs}
}
\value{
A bar plot
}
\description{
Visualize the importance of top 20 features
}
