% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{viz_reg_fit}
\alias{viz_reg_fit}
\title{Visualize regression model outputs}
\usage{
viz_reg_fit(models, names)
}
\arguments{
\item{models}{regression model outputs}

\item{names}{regression model names}
}
\value{
a point plot(s)
}
\description{
Visualize regression model outputs
}
