% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{viz_class_fit}
\alias{viz_class_fit}
\title{Visualize classification model outputs}
\usage{
viz_class_fit(model)
}
\arguments{
\item{models}{classification model outputs}

\item{names}{classification model names}
}
\value{
a bar plot(s)
}
\description{
Visualize classification model outputs
}
