% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{viz_class_fit}
\alias{viz_class_fit}
\title{Visualize a classification model output}
\usage{
viz_class_fit(
  model,
  model_title,
  test_x_class,
  test_y_class,
  metric_type = "class"
)
}
\arguments{
\item{model}{A classification model output}

\item{model_title}{A classification model name}

\item{test_x_class}{Outcomes of the test dataset}

\item{test_y_class}{Predictors of the test dataset}

\item{metric_type}{A type of the metrics ("class" or "probability"). Either a "class"-based metrics (e.g., accuracy, balanced accuracy, F-score) or a "probability"-based metrics (e.g., ROC Curve). The default type is "class."}
}
\value{
a bar or line plot(s)
}
\description{
Visualize a classification model output
}
